# reports/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Ashveil Console API",
    "version": "0.27.0",
    "description": "Generated API contract from Ashveil Express routes."
  },
  "servers": [
    {
      "url": "http://localhost:4160",
      "description": "Local backend server"
    }
  ],
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "paths": {
    "/api/access/menus": {
      "get": {
        "operationId": "getApiAccessMenus",
        "summary": "GET /api/access/menus",
        "tags": ["api/access"],
        "description": "Discovered from backend/src/server.js:240. Auth boundary: JWT bearer token required.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token"
          }
        },
        "x-source-file": "backend/src/server.js",
        "x-source-line": 240,
        "x-auth-boundary": "protected"
      }
    },
    "/api/access/permission-matrix": {
      "get": {
        "operationId": "getApiAccessPermissionMatrix",
        "summary": "GET /api/access/permission-matrix",
        "tags": ["api/access"],
        "description": "Discovered from backend/src/server.js:241. Auth boundary: JWT bearer token required.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token"
          }
        },
        "x-source-file": "backend/src/server.js",
        "x-source-line": 241,
        "x-auth-boundary": "protected"
      }
    },
    "/api/access/roles": {
      "get": {
        "operationId": "getApiAccessRoles",
        "summary": "GET /api/access/roles",
        "tags": ["api/access"],
        "description": "Discovered from backend/src/server.js:239. Auth boundary: JWT bearer token required.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token"
          }
        },
        "x-source-file": "backend/src/server.js",
        "x-source-line": 239,
        "x-auth-boundary": "protected"
      }
    },
    "/api/access/users": {
      "get": {
        "operationId": "getApiAccessUsers",
        "summary": "GET /api/access/users",
        "tags": ["api/access"],
        "description": "Discovered from backend/src/server.js:238. Auth boundary: JWT bearer token required.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token"
          }
        },
        "x-source-file": "backend/src/server.js",
        "x-source-line": 238,
        "x-auth-boundary": "protected"
      }
    },
    "/api/audit/logs": {
      "get": {
        "operationId": "getApiAuditLogs",
        "summary": "GET /api/audit/logs",
        "tags": ["api/audit"],
        "description": "Discovered from backend/src/server.js:246. Auth boundary: JWT bearer token required.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token"
          }
        },
        "x-source-file": "backend/src/server.js",
        "x-source-line": 246,
        "x-auth-boundary": "protected"
      }
    },
    "/api/audit/summary": {
      "get": {
        "operationId": "getApiAuditSummary",
        "summary": "GET /api/audit/summary",
        "tags": ["api/audit"],
        "description": "Discovered from backend/src/server.js:247. Auth boundary: JWT bearer token required.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token"
          }
        },
        "x-source-file": "backend/src/server.js",
        "x-source-line": 247,
        "x-auth-boundary": "protected"
      }
    },
    "/api/auth/login": {
      "post": {
        "operationId": "postApiAuthLogin",
        "summary": "POST /api/auth/login",
        "tags": ["api/auth"],
        "description": "Discovered from backend/src/server.js:206. Auth boundary: public endpoint.",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["username", "password"],
                "properties": {
                  "username": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 120
                  },
                  "password": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 256
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Invalid login payload"
          },
          "401": {
            "description": "Invalid username or password"
          }
        },
        "x-source-file": "backend/src/server.js",
        "x-source-line": 206,
        "x-auth-boundary": "public"
      }
    },
    "/api/auth/me": {
      "get": {
        "operationId": "getApiAuthMe",
        "summary": "GET /api/auth/me",
        "tags": ["api/auth"],
        "description": "Discovered from backend/src/server.js:219. Auth boundary: JWT bearer token required.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token"
          }
        },
        "x-source-file": "backend/src/server.js",
        "x-source-line": 219,
        "x-auth-boundary": "protected"
      }
    },
    "/api/dashboard": {
      "get": {
        "operationId": "getApiDashboard",
        "summary": "GET /api/dashboard",
        "tags": ["api/dashboard"],
        "description": "Discovered from backend/src/server.js:225. Auth boundary: JWT bearer token required.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token"
          }
        },
        "x-source-file": "backend/src/server.js",
        "x-source-line": 225,
        "x-auth-boundary": "protected"
      }
    },
    "/api/dictionaries": {
      "get": {
        "operationId": "getApiDictionaries",
        "summary": "GET /api/dictionaries",
        "tags": ["api/dictionaries"],
        "description": "Discovered from backend/src/server.js:245. Auth boundary: JWT bearer token required.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token"
          }
        },
        "x-source-file": "backend/src/server.js",
        "x-source-line": 245,
        "x-auth-boundary": "protected"
      }
    },
    "/api/health": {
      "get": {
        "operationId": "getApiHealth",
        "summary": "GET /api/health",
        "tags": ["api/health"],
        "description": "Discovered from backend/src/server.js:202. Auth boundary: public endpoint.",
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          }
        },
        "x-source-file": "backend/src/server.js",
        "x-source-line": 202,
        "x-auth-boundary": "public"
      }
    },
    "/api/ideas": {
      "get": {
        "operationId": "getApiIdeas",
        "summary": "GET /api/ideas",
        "tags": ["api/ideas"],
        "description": "Discovered from backend/src/server.js:291. Auth boundary: JWT bearer token required.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token"
          }
        },
        "x-source-file": "backend/src/server.js",
        "x-source-line": 291,
        "x-auth-boundary": "protected"
      }
    },
    "/api/jobs": {
      "get": {
        "operationId": "getApiJobs",
        "summary": "GET /api/jobs",
        "tags": ["api/jobs"],
        "description": "Discovered from backend/src/server.js:266. Auth boundary: JWT bearer token required.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token"
          }
        },
        "x-source-file": "backend/src/server.js",
        "x-source-line": 266,
        "x-auth-boundary": "protected"
      }
    },
    "/api/monitor": {
      "get": {
        "operationId": "getApiMonitor",
        "summary": "GET /api/monitor",
        "tags": ["api/monitor"],
        "description": "Discovered from backend/src/server.js:267. Auth boundary: JWT bearer token required.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token"
          }
        },
        "x-source-file": "backend/src/server.js",
        "x-source-line": 267,
        "x-auth-boundary": "protected"
      }
    },
    "/api/risk/events": {
      "get": {
        "operationId": "getApiRiskEvents",
        "summary": "GET /api/risk/events",
        "tags": ["api/risk"],
        "description": "Discovered from backend/src/server.js:271. Auth boundary: JWT bearer token required.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token"
          }
        },
        "x-source-file": "backend/src/server.js",
        "x-source-line": 271,
        "x-auth-boundary": "protected"
      }
    },
    "/api/risk/events/{eventKey}/status": {
      "patch": {
        "operationId": "patchApiRiskEventsByEventKeyStatus",
        "summary": "PATCH /api/risk/events/:eventKey/status",
        "tags": ["api/risk"],
        "description": "Discovered from backend/src/server.js:279. Auth boundary: JWT bearer token required.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [
          {
            "name": "eventKey",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "minLength": 1,
              "maxLength": 80,
              "pattern": "^[A-Za-z0-9:_-]+$"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["status"],
                "properties": {
                  "status": {
                    "type": "string",
                    "enum": ["pending", "processing", "confirmed", "ignored", "archived"]
                  },
                  "note": {
                    "type": "string",
                    "maxLength": 500
                  }
                },
                "additionalProperties": false
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "400": {
            "description": "Invalid risk event status payload or unsupported status"
          },
          "401": {
            "description": "Missing or invalid bearer token"
          },
          "404": {
            "description": "Risk event not found"
          },
          "503": {
            "description": "Prisma is unavailable for persisted risk event status updates"
          }
        },
        "x-source-file": "backend/src/server.js",
        "x-source-line": 279,
        "x-auth-boundary": "protected"
      }
    },
    "/api/risk/scores": {
      "get": {
        "operationId": "getApiRiskScores",
        "summary": "GET /api/risk/scores",
        "tags": ["api/risk"],
        "description": "Discovered from backend/src/server.js:268. Auth boundary: JWT bearer token required.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token"
          }
        },
        "x-source-file": "backend/src/server.js",
        "x-source-line": 268,
        "x-auth-boundary": "protected"
      }
    },
    "/api/watch/night": {
      "get": {
        "operationId": "getApiWatchNight",
        "summary": "GET /api/watch/night",
        "tags": ["api/watch"],
        "description": "Discovered from backend/src/server.js:292. Auth boundary: JWT bearer token required.",
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "parameters": [],
        "responses": {
          "200": {
            "description": "Successful response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "additionalProperties": true
                }
              }
            }
          },
          "401": {
            "description": "Missing or invalid bearer token"
          }
        },
        "x-source-file": "backend/src/server.js",
        "x-source-line": 292,
        "x-auth-boundary": "protected"
      }
    }
  },
  "x-generated-from": "reports/api-surface.json",
  "x-report": {
    "reportType": "ashveil_openapi_contract",
    "generatedAt": "2026-07-19T05:11:00.786Z",
    "ok": true,
    "summary": {
      "pathCount": 18,
      "operationCount": 18,
      "protectedOperationCount": 16,
      "publicOperationCount": 2
    },
    "gates": [
      {
        "name": "OpenAPI version",
        "ok": true,
        "detail": "3.1.0"
      },
      {
        "name": "API info version matches package",
        "ok": true,
        "detail": "0.27.0 / 0.27.0"
      },
      {
        "name": "generated from API surface report",
        "ok": true,
        "detail": "reports/api-surface.json"
      },
      {
        "name": "operation count matches API surface",
        "ok": true,
        "detail": "18/18"
      },
      {
        "name": "operation source lines recorded",
        "ok": true,
        "detail": "18/18 operation source line(s)"
      },
      {
        "name": "operation sources match API surface",
        "ok": true,
        "detail": "18 operation source mapping(s)"
      },
      {
        "name": "bearer security scheme present",
        "ok": true,
        "detail": "components.securitySchemes.bearerAuth"
      },
      {
        "name": "operation auth boundaries recorded",
        "ok": true,
        "detail": "18/18 operation auth boundary marker(s)"
      },
      {
        "name": "operation auth boundaries match API surface",
        "ok": true,
        "detail": "18 operation auth boundary mapping(s)"
      },
      {
        "name": "protected operations require bearer auth",
        "ok": true,
        "detail": "16 protected operations"
      },
      {
        "name": "public operations omit bearer auth",
        "ok": true,
        "detail": "2 public operations"
      },
      {
        "name": "server URL matches backend default port",
        "ok": true,
        "detail": "http://localhost:4160"
      },
      {
        "name": "Express path params converted",
        "ok": true,
        "detail": "colon params converted to {param}"
      },
      {
        "name": "operation IDs are unique",
        "ok": true,
        "detail": "18/18 unique operationId(s)"
      },
      {
        "name": "path parameters documented",
        "ok": true,
        "detail": "all templated path params documented"
      },
      {
        "name": "login request body constrained",
        "ok": true,
        "detail": "username<=120; password<=256; username/password only"
      },
      {
        "name": "login error responses documented",
        "ok": true,
        "detail": "400, 401"
      },
      {
        "name": "risk status request body constrained",
        "ok": true,
        "detail": "pending, processing, confirmed, ignored, archived; note<=500"
      },
      {
        "name": "risk event key path parameter constrained",
        "ok": true,
        "detail": "eventKey<=80; pattern=^[A-Za-z0-9:_-]+$"
      },
      {
        "name": "risk status error responses documented",
        "ok": true,
        "detail": "400, 404, 503"
      }
    ],
    "referenceBasis": [
      "OpenAPI Specification 3.1 contract document",
      "Bearer authentication boundary expressed as securitySchemes",
      "Generated from local Express route inventory, not hand-maintained text"
    ]
  }
}